+++ src/main/java/BOJ1431/solution.java
@@ -1,0 +1,2 @@
+package BOJ1431;public class solution {
+}
